Utilities for clustering

MSRDPReefs.specgram_scatter — Method

specgram_scatter(x, y, data; fs=1, nfft=256, noverlap=128, kwargs...)

Plot a scatter plot with each point is a spectrogram of a signal from data.

Examples:

julia> using MSRDPReefs

julia> using AbstractPlotting, GLMakie

julia> data = [randn(4800) for i ∈ 1:20];

julia> X = randn(2, length(data));

julia> specgram_scatter(X[1,:], X[2,:], data; fs=9600)
GLMakie.Screen(...)

MSRDPReefs.specgram_scatter — Method

specgram_scatter(x, y, data; fs=1, nfft=256, noverlap=128, kwargs...)

Plot a scatter plot with each point is a spectrogram of a WAV file from data.

Examples:

julia> using MSRDPReefs, WAV

julia> using AbstractPlotting, GLMakie

julia> root = mktempdir();

julia> [wavwrite(randn(4800), joinpath(root, string(i)*".wav"); Fs=9600) for i ∈ 1:20];

julia> wavpaths = readdir(root, join=true);

julia> X = randn(2, length(wavpaths));

julia> specgram_scatter(X[1,:], X[2,:], wavpaths; fs=9600)
GLMakie.Screen(...)

MSRDPReefs.visualizeclusters — Method

visualizeclusters(wavpaths, clusterindices, outlierindices, ndisplayperclass)

Display spectrograms of sounds in wavpaths according to the clustering result denoted by clusterindices and outlierindices. ndisplayperclass defines number of spectrogram per class to be displayed. The axes are blank if ndisplayperclass is larger than cluster size. When a mouse is hovering at thecenter of a spectrogram, the respective sound is played.

Example:

julia> using MSRDPReefs, WAV

julia> using AbstractPlotting, GLMakie

julia> ndata, nclass = 100, 3;

julia> root = mktempdir();

julia> data = randn(4800, ndata);

julia> [wavwrite(col, joinpath(root, string(i)*".wav"); Fs=9600) for (i, col) ∈ 
        enumerate(eachcol(data))];

julia> wavpaths = readdir(root, join=true);

julia> clusterindices, outlierindices = [1:9, 10:70, 71:91], 92:ndata;

julia> visualizeclusters(wavpaths, clusterindices, outlierindices, 10)
GLMakie.Screen(...)

MSRDPReefs.visualizeclusters — Method

visualizeclusters(wavpaths, res, ndisplayperclass)

Display spectrograms of sounds in wavpaths according to the clustering result res. ndisplayperclass defines number of spectrogram per class to be displayed. The axes are blank if ndisplayperclass is larger than cluster size. When a mouse is hovering at thecenter of a spectrogram, the respective sound is played.

Examples:

julia> using Clustering, MSRDPReefs, WAV

julia> using AbstractPlotting, GLMakie

julia> ndata, nclass = 100, 3;

julia> root = mktempdir();

julia> data = randn(4800, ndata);

julia> [wavwrite(col, joinpath(root, string(i)*".wav"); Fs=9600) for (i, col) ∈ 
        enumerate(eachcol(data))];

julia> wavpaths = readdir(root, join=true);

julia> res = kmeans(data, nclass);

julia> visualizeclusters(wavpaths, res, 10)
GLMakie.Screen(...)