Data loading

MSRDPReefs.SITEDTRANGES — Constant

Default datetime ranges for all the sites.

Example:

julia> SITEDTRANGES
Dict{String,Array{Tuple{Dates.DateTime,Dates.DateTime},1}} with 10 entries:
  "TPT"        => [(DateTime("2019-11-22T11:30:00"), DateTime("2019-12-31T00:00:00")), (DateTime("202…
  "RL-W"       => [(DateTime("2019-06-21T13:00:00"), DateTime("2019-07-15T11:00:00")), (DateTime("202…
  "Semakau-SW" => [(DateTime("2019-10-11T14:00:00"), DateTime("2019-10-17T03:00:00")), (DateTime("202…
  "SL-SE"      => [(DateTime("2019-10-11T12:00:00"), DateTime("2019-10-25T00:00:00")), (DateTime("202…
  "Seringat"   => [(DateTime("2019-10-11T11:00:00"), DateTime("2019-11-08T00:00:00")), (DateTime("202…
  "Kusu-NE"    => [(DateTime("2019-09-11T11:00:00"), DateTime("2019-10-11T10:30:00")), (DateTime("202…
  "Hantu-W"    => [(DateTime("2019-09-11T13:00:00"), DateTime("2019-10-11T12:00:00")), (DateTime("202…
  "Jong-S"     => [(DateTime("2019-09-11T12:30:00"), DateTime("2019-10-11T12:00:00")), (DateTime("202…
  "Semakau-NW" => [(DateTime("2020-01-08T13:30:00"), DateTime("2020-02-16T00:00:00")), (DateTime("202…
  "SD-NW"      => [(DateTime("2019-11-22T14:00:00"), DateTime("2019-12-02T11:30:00")), (DateTime("202…

MSRDPReefs._metadata — Function

_metadata(path, 
          dtranges=SITEDTRANGES[split(path, "/")[end]], 
          divinglogpath::AbstractString=joinpath(path, "../logs/diving-log.csv")

Get metadata in DataFrame of data collected at a particular site.

Datetime

Site

Sensitivity

Gain

Wavpath

Battery

Depth

Temperature

Red light

Blue light

Green light

Diver

Example:

julia> df = MSRDPReefs._metadata("/home/arl/Data/reefwatch/recordings/Hantu-W");

julia> first(df, 5)
5×13 DataFrame
│ Row │ Datetime            │ Site    │ Sensitivity │ Gain    │ Wavpath                                                                         │ Battery  │ Depth    │ Temperature │ Redlight │ Greenlight │ Bluelight │ Moonphase │ Diver │
│     │ Dates.DateTime      │ String  │ Float64     │ Float64 │ String                                                                          │ Float64? │ Float64? │ Float64?    │ Float64? │ Float64?   │ Float64?  │ Int64     │ Bool  │
├─────┼─────────────────────┼─────────┼─────────────┼─────────┼─────────────────────────────────────────────────────────────────────────────────┼──────────┼──────────┼─────────────┼──────────┼────────────┼───────────┼───────────┼───────┤
│ 1   │ 2019-09-11T13:02:08 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T130208.wav │ 4.31     │ 2.61     │ 30.3        │ 6303.0   │ 10794.0    │ 6907.0    │ 5         │ 0     │
│ 2   │ 2019-09-11T13:07:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T130707.wav │ 4.3      │ 2.6      │ 30.28       │ 6911.0   │ 10755.0    │ 7262.0    │ 5         │ 0     │
│ 3   │ 2019-09-11T13:12:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T131207.wav │ 4.3      │ 2.56     │ 30.21       │ 6407.0   │ 10623.0    │ 6929.0    │ 5         │ 0     │
│ 4   │ 2019-09-11T13:17:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T131707.wav │ 4.28     │ 2.54     │ 30.22       │ 6008.0   │ 10385.0    │ 6814.0    │ 5         │ 0     │
│ 5   │ 2019-09-11T13:22:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T132207.wav │ 4.28     │ 2.51     │ 30.23       │ 6227.0   │ 10614.0    │ 6899.0    │ 5         │ 0     │

MSRDPReefs.datacollectionprogress — Method

datacollectionprogress(paths)

Show the data collection progress in percentage given multipe parent paths. 100 percent equals to one month of data.

Examples:

julia> datacollectionprogress(["/home/arl/Data/reefwatch/recordings/Hantu-W",
"/home/arl/Data/reefwatch/recordings/RL-W"])
2×7 DataFrame
│ Row │ Site    │ D1 (%)  │ D2 (%)  │ D3 (%)  │ D4 (%)  │ D5 (%)  │ D6 (%)  │
│     │ String  │ Real?   │ Real?   │ Real?   │ Real?   │ Real?   │ Real?   │
├─────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ 1   │ Hantu-W │ 100.788 │ 128.408 │ 77.8376 │ 103.189 │ missing │ missing │
│ 2   │ RL-W    │ 83.6276 │ 128.501 │ 116.384 │ missing │ missing │ missing │

MSRDPReefs.datacollectionprogress — Method

datacollectionprogress(rootpath)

Show the data collection progress in percentage given a root path. 100 percent equals to one month of data.

Examples:

julia> datacollectionprogress("/home/arl/Data/reefwatch/recordings")
10×7 DataFrame
│ Row │ Site       │ D1 (%)  │ D2 (%)  │ D3 (%)  │ D4 (%)  │ D5 (%)  │ D6 (%)  │
│     │ String     │ Real?   │ Real?   │ Real?   │ Real?   │ Real?   │ Real?   │
├─────┼────────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ 1   │ Hantu-W    │ 100.788 │ 128.408 │ 77.8376 │ 103.189 │ missing │ missing │
│ 2   │ Jong-S     │ 100.811 │ 116.446 │ 103.188 │ missing │ missing │ missing │
│ 3   │ Kusu-NE    │ 100.452 │ 116.518 │ 103.32  │ missing │ missing │ missing │
│ 4   │ RL-W       │ 83.6276 │ 128.501 │ 116.384 │ 103.324 │ missing │ missing │
│ 5   │ SD-NW      │ 33.1604 │ 44.6388 │ 103.184 │ 103.181 │ missing │ missing │
│ 6   │ SL-SE      │ 44.5767 │ 103.188 │ missing │ missing │ missing │ missing │
│ 7   │ Semakau-NW │ 128.225 │ 116.374 │ 103.11  │ missing │ missing │ missing │
│ 8   │ Semakau-SW │ 18.9137 │ 4.374   │ 103.383 │ missing │ missing │ missing │
│ 9   │ Seringat   │ 90.5111 │ 51.5231 │ 106.451 │ 103.187 │ missing │ missing │
│ 10  │ TPT        │ 128.354 │ 25.6559 │ 106.307 │ 103.043 │ missing │ missing │

MSRDPReefs.getlogpaths — Method

getlogpaths(path, dtranges)

Get paths of all the log files in parent path path

MSRDPReefs.getpaths — Method

getpaths(rootpath)

Get parent paths. The output is a 1-D array containing paths for data (acoustic and environmental) collected at all the sites.

MSRDPReefs.metadata — Function

metadata(rootpath, dtrangesdict=SITEDTRANGES)

Get metadata in DataFrame of data collected at all the sites.

Datetime

Site

Sensitivity

Gain

Wavpath

Battery

Depth

Temperature

Red light

Blue light

Green light

Diver

Example:

julia> df = metadata("/home/arl/Data/reefwatch/recordings");

julia> first(df, 5)
5×13 DataFrame
│ Row │ Datetime            │ Site    │ Sensitivity │ Gain    │ Wavpath                                                                         │ Battery  │ Depth    │ Temperature │ Redlight │ Greenlight │ Bluelight │ Moonphase │ Diver │
│     │ Dates.DateTime      │ String  │ Float64     │ Float64 │ String                                                                          │ Float64? │ Float64? │ Float64?    │ Float64? │ Float64?   │ Float64?  │ Int64     │ Bool  │
├─────┼─────────────────────┼─────────┼─────────────┼─────────┼─────────────────────────────────────────────────────────────────────────────────┼──────────┼──────────┼─────────────┼──────────┼────────────┼───────────┼───────────┼───────┤
│ 1   │ 2019-09-11T13:02:08 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T130208.wav │ 4.31     │ 2.61     │ 30.3        │ 6303.0   │ 10794.0    │ 6907.0    │ 5         │ 0     │
│ 2   │ 2019-09-11T13:07:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T130707.wav │ 4.3      │ 2.6      │ 30.28       │ 6911.0   │ 10755.0    │ 7262.0    │ 5         │ 0     │
│ 3   │ 2019-09-11T13:12:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T131207.wav │ 4.3      │ 2.56     │ 30.21       │ 6407.0   │ 10623.0    │ 6929.0    │ 5         │ 0     │
│ 4   │ 2019-09-11T13:17:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T131707.wav │ 4.28     │ 2.54     │ 30.22       │ 6008.0   │ 10385.0    │ 6814.0    │ 5         │ 0     │
│ 5   │ 2019-09-11T13:22:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T132207.wav │ 4.28     │ 2.51     │ 30.23       │ 6227.0   │ 10614.0    │ 6899.0    │ 5         │ 0     │

julia> unique(df.Site)
10-element Array{String,1}:
 "Hantu-W"
 "Jong-S"
 "Kusu-NE"
 "RL-W"
 "SD-NW"
 "SL-SE"
 "Semakau-NW"
 "Semakau-SW"
 "Seringat"
 "TPT"

MSRDPReefs.metadata — Function

metadata(paths, dtrangesdict=SITEDTRANGES)

Get metadata in DataFrame of data collected at multiple sites.

Datetime

Site

Sensitivity

Gain

Wavpath

Battery

Depth

Temperature

Red light

Blue light

Green light

Diver

Example:

julia> df = metadata(["/home/arl/Data/reefwatch/recordings/Hantu-W","/home/arl/Data/reefwatch/recordings/RL-W"]);

julia> first(df, 5)
5×13 DataFrame
│ Row │ Datetime            │ Site    │ Sensitivity │ Gain    │ Wavpath                                                                         │ Battery  │ Depth    │ Temperature │ Redlight │ Greenlight │ Bluelight │ Moonphase │ Diver │
│     │ Dates.DateTime      │ String  │ Float64     │ Float64 │ String                                                                          │ Float64? │ Float64? │ Float64?    │ Float64? │ Float64?   │ Float64?  │ Int64     │ Bool  │
├─────┼─────────────────────┼─────────┼─────────────┼─────────┼─────────────────────────────────────────────────────────────────────────────────┼──────────┼──────────┼─────────────┼──────────┼────────────┼───────────┼───────────┼───────┤
│ 1   │ 2019-09-11T13:02:08 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T130208.wav │ 4.31     │ 2.61     │ 30.3        │ 6303.0   │ 10794.0    │ 6907.0    │ 5         │ 0     │
│ 2   │ 2019-09-11T13:07:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T130707.wav │ 4.3      │ 2.6      │ 30.28       │ 6911.0   │ 10755.0    │ 7262.0    │ 5         │ 0     │
│ 3   │ 2019-09-11T13:12:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T131207.wav │ 4.3      │ 2.56     │ 30.21       │ 6407.0   │ 10623.0    │ 6929.0    │ 5         │ 0     │
│ 4   │ 2019-09-11T13:17:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T131707.wav │ 4.28     │ 2.54     │ 30.22       │ 6008.0   │ 10385.0    │ 6814.0    │ 5         │ 0     │
│ 5   │ 2019-09-11T13:22:07 │ Hantu-W │ -179.8      │ 16.77   │ /home/arl/Data/reefwatch/recordings/Hantu-W/2019-09/2019-09/20190911T132207.wav │ 4.28     │ 2.51     │ 30.23       │ 6227.0   │ 10614.0    │ 6899.0    │ 5         │ 0     │

julia> last(df, 5)
5×13 DataFrame
│ Row │ Datetime            │ Site   │ Sensitivity │ Gain    │ Wavpath                                                                      │ Battery  │ Depth    │ Temperature │ Redlight │ Greenlight │ Bluelight │ Moonphase │ Diver │
│     │ Dates.DateTime      │ String │ Float64     │ Float64 │ String                                                                       │ Float64? │ Float64? │ Float64?    │ Float64? │ Float64?   │ Float64?  │ Int64     │ Bool  │
├─────┼─────────────────────┼────────┼─────────────┼─────────┼──────────────────────────────────────────────────────────────────────────────┼──────────┼──────────┼─────────────┼──────────┼────────────┼───────────┼───────────┼───────┤
│ 1   │ 2020-09-11T08:39:15 │ RL-W   │ -179.9      │ 16.77   │ /home/arl/Data/reefwatch/recordings/RL-W/2020-08/2020-09/20200911T083915.wav │ 3.76     │ 2.43     │ 29.64       │ 282.0    │ 509.0      │ 283.0     │ 8         │ 0     │
│ 2   │ 2020-09-11T08:44:16 │ RL-W   │ -179.9      │ 16.77   │ /home/arl/Data/reefwatch/recordings/RL-W/2020-08/2020-09/20200911T084416.wav │ 3.77     │ 2.42     │ 29.66       │ 291.0    │ 527.0      │ 293.0     │ 8         │ 0     │
│ 3   │ 2020-09-11T08:49:17 │ RL-W   │ -179.9      │ 16.77   │ /home/arl/Data/reefwatch/recordings/RL-W/2020-08/2020-09/20200911T084917.wav │ 3.77     │ 2.41     │ 29.65       │ 297.0    │ 548.0      │ 310.0     │ 8         │ 0     │
│ 4   │ 2020-09-11T08:54:18 │ RL-W   │ -179.9      │ 16.77   │ /home/arl/Data/reefwatch/recordings/RL-W/2020-08/2020-09/20200911T085418.wav │ 3.78     │ 2.42     │ 29.65       │ 322.0    │ 594.0      │ 331.0     │ 8         │ 0     │
│ 5   │ 2020-09-11T08:59:19 │ RL-W   │ -179.9      │ 16.77   │ /home/arl/Data/reefwatch/recordings/RL-W/2020-08/2020-09/20200911T085919.wav │ 3.79     │ 2.39     │ 29.62       │ 342.0    │ 634.0      │ 352.0     │ 8         │ 0     │

MSRDPReefs.revisedf! — Method

revisedf!(df, logpath, dtrange, acousticsensor)

Modify the dataframe df by adding Datetime, Site, Sensitivity, Gain, Wawpath, Moonphase, Diver columns, combining Red, Green and Blue into RGB(Red, Green, Blue), renaming the columns and removing rows which are not in the datetime range.