Utilities for clustering

MSRDPReefs.visualizeclusters — Method

visualizeclusters(wavpaths, clusterindices, outlierindices, ndisplayperclass)

Display spectrograms of sounds in wavpaths according to the clustering result denoted by clusterindices and outlierindices. ndisplayperclass defines number of spectrogram per class to be displayed. The axes are blank if ndisplayperclass is larger than cluster size. When a mouse is hovering at thecenter of a spectrogram, the respective sound is played.

Example:

julia> using MSRDPReefs, WAV

julia> using AbstractPlotting, GLMakie

julia> ndata, nclass = 100, 3;

julia> root = mktempdir();

julia> data = randn(4800, ndata);

julia> [wavwrite(col, joinpath(root, string(i)*".wav"); Fs=9600) for (i, col) ∈ 
        enumerate(eachcol(data))];

julia> wavpaths = readdir(root, join=true);

julia> clusterindices, outlierindices = [1:9, 10:70, 71:91], 92:ndata;

julia> visualizeclusters(wavpaths, clusterindices, outlierindices, 10)
GLMakie.Screen(...)

MSRDPReefs.visualizeclusters — Method

visualizeclusters(wavpaths, res, ndisplayperclass)

Display spectrograms of sounds in wavpaths according to the clustering result res. ndisplayperclass defines number of spectrogram per class to be displayed. The axes are blank if ndisplayperclass is larger than cluster size. When a mouse is hovering at thecenter of a spectrogram, the respective sound is played.

Examples:

julia> using Clustering, MSRDPReefs, WAV

julia> using AbstractPlotting, GLMakie

julia> ndata, nclass = 100, 3;

julia> root = mktempdir();

julia> data = randn(4800, ndata);

julia> [wavwrite(col, joinpath(root, string(i)*".wav"); Fs=9600) for (i, col) ∈ 
        enumerate(eachcol(data))];

julia> wavpaths = readdir(root, join=true);

julia> res = kmeans(data, nclass);

julia> visualizeclusters(wavpaths, res, 10)
GLMakie.Screen(...)